// src/configs/enums/main.enum.ts
export enum MainEnum {
  APP_TOKEN = "SYNERIS_SERVICE_TOKEN",
  USER_NAME = "SYNERIS_SERVICE_USER_NAME",
  USER_ID = "SYNERIS_SERVICE_USER_ID",
  ROLE_NAME = "SYNERIS_SERVICE_ROLE_NAME",
  ROLE_ID = "SYNERIS_SERVICE_ROLE_ID",
  EMPLOYEE_ID = "SYNERIS_SERVICE_EMPLOYEE_ID",
  ACCESS_ARRAY = "SYNERIS_SERVICE_ACCESS_ARRAY",
}